Rename typo
