+++ client/src/app/core/helpers/pagination-helper.ts
